+++ sqlalchemy_bigquery/__init__.py
@@ -16,3 +16,21 @@
 # COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 # IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 # CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+from base import BigQueryDialect
+from base import (
+    STRING,
+    BOOL,
+    BOOLEAN,
+    INT64,
+    INTEGER,
+    FLOAT64,
+    FLOAT,
+    TIMESTAMP,
+    DATETIME,
+    DATE,
+    BYTES,
+    TIME,
+    RECORD,
+    NUMERIC,
+    BIGNUMERIC,
+    )
