+++ tests/system/test_sqlalchemy_bigquery.py
@@ -22,7 +22,7 @@ from __future__ import unicode_literals
 
 from google.api_core.exceptions import BadRequest
 from sqlalchemy_bigquery.api import ApiClient
-from sqlalchemy_bigquery.sqlalchemy_bigquery import BigQueryDialect
+from sqlalchemy_bigquery import BigQueryDialect
 from sqlalchemy.engine import create_engine
 from sqlalchemy.schema import Table, MetaData, Column
 from sqlalchemy.ext.declarative import declarative_base
